machine-learning-client/requirements.txt
opencv-python==4.9.0.80
torch==2.2.2
torchvision==0.17.2
pymongo==4.6.3
Pillow==9.0.1